PODS-1392 Moved Are you in UK page to correct place
