+++ app/views/register/areYouInUK.scala.html
@@ -15,7 +15,7 @@
  *@
 
 @import config.FrontendAppConfig
-@import controllers.register.company.routes._
+@import controllers.register.routes._
 @import uk.gov.hmrc.play.views.html._
 
 @(appConfig: FrontendAppConfig, form: Form[_])(implicit request: Request[_], messages: Messages)
