+++ conf/company.routes
@@ -83,6 +83,3 @@ GET        /select-contact-address                      controllers.register.com
 POST       /select-contact-address                      controllers.register.company.CompanyContactAddressListController.onSubmit(mode: Mode = NormalMode)
 GET        /change-select-contact-address               controllers.register.company.CompanyContactAddressListController.onPageLoad(mode: Mode = CheckMode)
 POST       /change-select-contact-address               controllers.register.company.CompanyContactAddressListController.onSubmit(mode: Mode = CheckMode)
-
-GET        /country-of-work                             controllers.register.company.AreYouInUKController.onPageLoad()
-POST       /country-of-work                             controllers.register.company.AreYouInUKController.onSubmit()
